[chore]: change name board -> post
